+++ lib/public/StormByte/wcstring.cxx
@@ -1,0 +1,82 @@
+/*
+ * Copyright (C) 2024-2026 David C. Manuelda (StormBytePP)
+ *
+ * This file is part of StormByte.
+ *
+ * StormByte is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License version 3
+ * or later, as published by the Free Software Foundation.
+ *
+ * StormByte is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with StormByte. If not, see
+ * <https://www.gnu.org/licenses/lgpl-3.0.html>.
+ */
+
+#include <StormByte/wcstring.hxx>
+
+#include <cwchar>
+
+using namespace StormByte;
+
+const wchar_t* WCString::Duplicate(const wchar_t* str) noexcept {
+	if (!str)
+		return nullptr;
+	const std::size_t len = std::wcslen(str) + 1;
+	wchar_t* out = new wchar_t[len];
+	std::wmemcpy(out, str, len);
+	return out;
+}
+
+WCString::WCString() noexcept
+: m_data(nullptr) {}
+
+WCString::WCString(const wchar_t* str) noexcept
+: m_data(Duplicate(str)) {}
+
+WCString::WCString(const WCString& other) noexcept
+: m_data(Duplicate(other.m_data)) {}
+
+WCString::WCString(WCString&& other) noexcept
+: m_data(other.m_data) {
+	other.m_data = nullptr;
+}
+
+WCString::~WCString() noexcept {
+	delete[] m_data;
+	m_data = nullptr;
+}
+
+WCString& WCString::operator=(const WCString& other) noexcept {
+	if (this != &other) {
+		delete[] m_data;
+		m_data = Duplicate(other.m_data);
+	}
+	return *this;
+}
+
+WCString& WCString::operator=(WCString&& other) noexcept {
+	if (this != &other) {
+		delete[] m_data;
+		m_data = other.m_data;
+		other.m_data = nullptr;
+	}
+	return *this;
+}
+
+void WCString::Reset(const wchar_t* str) noexcept {
+	delete[] m_data;
+	m_data = Duplicate(str);
+}
+
+std::size_t WCString::Length() const noexcept {
+	return m_data ? std::wcslen(m_data) : 0;
+}
+
+WCString::operator const wchar_t*() const noexcept {
+	return m_data;
+}